requirements.txt
Pillow>=10.0.0
requests>=2.20.0
opencv-python>=4.2.0.32
tqdm>=4.23.0
diffimg==0.2.3
arabic-reshaper==2.1.3
python-bidi==0.4.2
wikipedia>=1.4.0
diffimg
opencv-python
numpy
arabic-reshaper
python-bidi
wikipedia
tqdm